Compute inverted call tree timings differently.
This replaces lots of uses of getCallNodeTable() with uses of
getNonInvertedCallNodeTable().
It also replaces lots of uses of getStackIndexToCallNodeIndex() with
uses of getStackIndexToNonInvertedCallNodeIndex().

We now compute the call tree timings quite differently for inverted mode
compared to non-inverted mode. There's one part of the work that's shared:
The getCallNodeLeafAndSummary computes the self time for each non-inverted
node, and the result is used for both the inverted and the non-inverted
call tree timings.
The CallTreeTimings Flow type is turned into an enum, with a different
type for CallTreeTimingsNonInverted and for CallTreeTimingsInverted.
A new implementation for the CallTreeInternal interface is added.