Profile format changes for flows and stack-based markers
This changeset changes the type definition for marker schema
field types and for the marker schema itself, to include the new
field types 'flow-id' and 'terminating-flow-id'. It also has some
minimal changes to treat the new field types as if they were unique-string
fields.
Somewhat unrelated to flows, it also adds an optional `isStackBased`
field to the marker schema type, which is currently unused. This is
for https://github.com/firefox-devtools/profiler/issues/3141 .

Additionally, it increments the version of the two profile formats,
so that profiles with the new field types won't be loaded by
older front-ends.